+++ src/core/timerha.h
