+++ src/modules/rt_events/RT_E_CYCLE_fbt.h
@@ -24,7 +24,7 @@
 
 #include "core/esfb.h"
 #include "core/ecet.h"
-#include "arch/timerha.h"
+#include "core/timerha.h"
 
 class FORTE_RT_E_CYCLE final : public CEventSourceFB {
     DECLARE_FIRMWARE_FB(FORTE_RT_E_CYCLE)
@@ -33,7 +33,6 @@ class FORTE_RT_E_CYCLE final : public CEventSourceFB {
     static const TEventID scmEventSTARTID = 0;
     static const TEventID scmEventSTOPID = 1;
     static const TEventID scmEventEOID = 0;
-
 
     bool mActive; //!> flag to indicate that the timed fb is currently active
     CEventChainExecutionThread mECEO;
